requirements.txt
supabase==2.19.0
python-dotenv==1.0.1
pydantic==2.8.2
python-dateutil==2.9.0
pytest==8.3.2
pytest-asyncio==0.24.0
pandas==2.2.2
numpy==2.0.1
fastapi==0.112.1
uvicorn[standard]==0.30.6
pathlib2==2.3.7
humanize==4.10.0
colorlog==6.8.2
rich==13.8.0
cryptography==43.0.1
scikit-learn==1.5.1